Various code optimizations and cleanups:
* Use StringBuilder instead of StringBuffer
* Use pre-calculated array lengths whenever possible
* Remove unnecessary String constructor and String.toString() invocations
* Remove unnecessary new Integer, Long.valueOf, Integer.intValue, Boolean.valueOf invocations
* Replace one-letter String searches of indexOf / lastIndexOf with character literal
* Use isEmpty instead of String.equals(""), size() == 0 or similar
* Remove unnecessary new Class[] {} and new Object[] {}
* Use autoboxing instead of new Class[] {...} and new Object[] {...}

+++ core/containers/tomcat/src/main/java/org/codehaus/cargo/container/tomcat/internal/NonceCounter.java
@@ -70,15 +70,15 @@ public class NonceCounter
         Integer count = nonces.get(nonce);
         if (count == null)
         {
-            count = Integer.valueOf(1);
+            count = 1;
         }
         else
         {
-            count = Integer.valueOf(count.intValue() + 1);
+            count = count + 1;
         }
 
         nonces.put(nonce, count);
 
-        return String.format("%08x", count.intValue());
+        return String.format("%08x", count);
     }
 }
